+++ src/mdacli/__main__.py
@@ -19,7 +19,6 @@ Documentation for each module can be found at the respective sections on the
 .. _`MDAnalysis Documentation`:
    https://docs.mdanalysis.org/stable/documentation_pages/analysis_modules.html
 """
-
 from MDAnalysis.analysis import __all__
 
 import mdacli
@@ -38,7 +37,7 @@ def main():
                  'InterRDF_s']
 
     mdacli.cli(name="MDAnalysis",
-               module_list=__all__,
+               module_list=[f'MDAnalysis.analysis.{m}' for m in __all__],
                version=mdacli.__version__,
                description=__doc__,
                skip_modules=skip_mods,