+++ src/mdacli/save.py
@@ -17,7 +17,7 @@ import numpy as np
 from MDAnalysis.analysis.base import Results
 
 
-def save_results(results, fprefix="mdacli_results"):
+def save(results, fprefix="mdacli_results"):
     """
     Save the attributes of a results instance to disk.
 
@@ -277,7 +277,7 @@ def save_Results_object(results, fprefix='results', remove=True):
     keys = []
     for key, value in results.items():
         if isinstance(value, Results):
-            save_results(f"{fprefix}_{key}", value)
+            save(f"{fprefix}_{key}", value)
             keys.append(key)
 
     return return_with_remove(results, keys, remove)